ex01, need to rework algo, i doesn work for basicaly all 2 numbers followed by sign

+++ ex01/includes/RPN.hpp
@@ -5,7 +5,7 @@
 #include "colors.hpp"
 #include <stack>
 
-std::stack<int> fill_stack(std::string arg);
-double calcules_loop(std::string arg, std::stack<int> stack);
+std::stack<long> fill_stack(std::string arg);
+long calcules_loop(std::string arg, std::stack<long> stack);
 
 #endif
